+++ app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Update to `heapless-bytes` v0.5 and `trussed-core` v0.2
+
 ## [v0.1.0](https://github.com/trussed-dev/ctaphid-dispatch/releases/tag/app-v0.1.0) (2025-01-08)
 
 Initial release that extracts the `app` and `command` modules from `ctaphid-dispatch` 0.1 into a separate crate.