+++ dispatch/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
--
+- Update to `heapless-bytes` v0.5 and `trussed-core` v0.2
 
 ## [0.3.0] - 2025-03-21